+++ code/broker-server/src/main/java/com/google/cloud/broker/apps/brokerserver/accesstokens/AccessTokenCacheFetcher.java
@@ -18,6 +18,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.grpc.Status;
 import org.slf4j.MDC;
 
+import com.google.cloud.broker.apps.brokerserver.logging.LoggingUtils;
 import com.google.cloud.broker.apps.brokerserver.accesstokens.providers.AbstractProvider;
 import com.google.cloud.broker.apps.brokerserver.validation.Validation;
 import com.google.cloud.broker.caching.CacheFetcher;
@@ -63,7 +64,7 @@ public class AccessTokenCacheFetcher extends CacheFetcher {
         catch (IllegalArgumentException e) {
             throw Status.PERMISSION_DENIED.withDescription("Principal `" + owner + "` cannot be matched to a Google identity.").asRuntimeException();
         }
-        MDC.put("access_token_user", googleIdentity);
+        MDC.put(LoggingUtils.MDC_ACCESS_TOKEN_USER_KEY, googleIdentity);
         AccessToken accessToken = AbstractProvider.getInstance().getAccessToken(googleIdentity, scopes);
         return AccessBoundaryUtils.addAccessBoundary(accessToken, target);
     }