+++ code/core/src/main/java/com/google/cloud/broker/encryption/backends/DummyEncryptionBackend.java
@@ -11,6 +11,8 @@
 
 package com.google.cloud.broker.encryption.backends;
 
+import com.google.cloud.broker.checks.CheckResult;
+
 /**
  Dummy encryption backend that does not encrypt nor decrypt anything.
  Use only for testing. Do NOT use in production!
@@ -25,6 +27,11 @@ public class DummyEncryptionBackend extends AbstractEncryptionBackend {
     @Override
     public byte[] encrypt(byte[] plainText) {
         return plainText;
+    }
+
+    @Override
+    public CheckResult checkConnection() {
+        return new CheckResult(true);
     }
 
 }