+++ code/extensions/database/cloud-datastore/src/main/java/com/google/cloud/broker/database/backends/CloudDatastoreBackend.java
@@ -11,12 +11,15 @@
 
 package com.google.cloud.broker.database.backends;
 
+import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.util.*;
 
-import com.google.cloud.broker.database.models.Model;
-import com.google.cloud.broker.settings.AppSettings;
 import com.google.cloud.datastore.*;
 
+import com.google.cloud.broker.checks.CheckResult;
+import com.google.cloud.broker.database.models.Model;
+import com.google.cloud.broker.settings.AppSettings;
 import com.google.cloud.broker.database.DatabaseObjectNotFound;
 
 
@@ -120,7 +123,23 @@ public class CloudDatastoreBackend extends AbstractDatabaseBackend {
 
     @Override
     public void initializeDatabase() {
-        // Cloud Datastore doesn't need to do any initialization. A table is automatically be created
-        // when the first object is inserted.
+        // Cloud Datastore doesn't need to do any initialization.
+        // A table is automatically be created when the first object is inserted.
+    }
+
+    @Override
+    public CheckResult checkConnection() {
+        try {
+            Datastore datastore = getService();
+            Query<Entity> query = Query.newEntityQueryBuilder()
+                .setKind("ABCDEFGHIJ1234567890")  // Arbitrary fictitious Kind
+                .build();
+            datastore.run(query);
+            return new CheckResult(true);
+        } catch(Exception e) {
+            StringWriter sw = new StringWriter();
+            e.printStackTrace(new PrintWriter(sw));
+            return new CheckResult(false, sw.toString());
+        }
     }
 }