+++ CHANGES.md
@@ -1,3 +1,12 @@
+## 0.10.0 (March 20, 2020)
+
+- Disabled credential access boundary by default.
+- Added new base client library: `broker-client-lib`.
+- Renamed `broker-connector` package to `broker-hadoop-connector`.
+- Added `PingServer` command.
+- Added system check when the broker server starts to test connectivity
+  with database, cache, and encryption backends.
+
 ## 0.9.2 (March 9, 2020)
 
 - Added support for credentials access boundary.